backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.42.30
botocore==1.42.30
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.3.1
DateTime==4.5
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
df2gspread==1.0.4
entrypoints==0.4
exceptiongroup==1.3.1
executing==1.0.0
Faker==38.0.0
fastapi==0.128.0
fastjsonschema==2.16.1
google-api-python-client==1.6.7
google-auth==2.11.0
google-auth-oauthlib==0.5.2
gspread==5.5.0
h11==0.16.0
httplib2==0.20.4
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.10
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
panda==0.3.1
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.16.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.50.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==3.0.1
urllib3==1.26.12
uvicorn==0.40.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
zope.interface==5.4.0
